fixed clang warnings

+++ source/mpi/ProcessorUpdate.cpp
@@ -8,7 +8,7 @@ double ProcessorUpdate::GetLoadImbalance(Tessellation const& tlocal)const
 	int ws=get_mpi_size();
 	vector<int> N0(ws,0),N(ws,0);
 	int n=tlocal.GetPointNo();
-	N0[get_mpi_rank()]=n;
+	N0[static_cast<size_t>(get_mpi_rank())]=n;
 	MPI_Allgather(&n,1,MPI_INT,&N[0],1,MPI_INT,MPI_COMM_WORLD);
 	int total=0;
 	for(size_t i=0;i<N.size();++i)
